refactor: mapping 패키지 추가
